+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [0.0.3] - New property builder
+
+- Add support for gender and plural for [easy_localization](https://pub.dev/packages/easy_localization)
+- Add support for JSON keys with symbols
+- Add support for JSON keys with spaces
+- Add support for JSON keys with number as first char
+
 ## [0.0.2] - New default class name
 
 - Default class name moved to `TranslationsKeychain`